==> CIC/1_CIC_original_dataset.ipynb <==
# Read original data
import dask.dataframe as dd
import numpy as np
import pandas as pd
from sklearn.utils import shuffle

df = pd.read_csv("../dataset/CIC-2018/original.csv")
print('Original dataset:')
print(df['Label'].value_counts())
print("----------------------------------")
df['split'] = np.random.randn(df.shape[0], 1)

msk = np.random.rand(len(df)) <= 0.7

train = df[msk]
test = df[~msk]
train = train.drop(['split'],axis=1)
test = test.drop(['split'],axis=1)

#train.to_csv("../dataset/CIC-2018/train_ori.csv", index=False, header = True)
#test.to_csv("../dataset/CIC-2018/test.csv", index=False, header = True)
print(train['Label'].value_counts())
print("----------------------------------")
print(test['Label'].value_counts())

import numpy as np
import pandas as pd
from sklearn.utils import shuffle
import warnings
from fastprogress import fastprogress
warnings.filterwarnings("ignore")

dt = pd.read_csv("../dataset/CIC-2018/train_ori.csv")
print(dt['Label'].value_counts())
# Random data attack and background 
attack1 = dt[dt['Label']=="DoS attacks-SlowHTTPTest"]
attack2 = dt[dt['Label']=="DoS attacks-Slowloris"]
attack3 = dt[dt['Label']=="DDOS attack-LOIC-UDP"]
attack4 = dt[dt['Label']=="Brute Force -Web"]
attack5 = dt[dt['Label']=="Brute Force -XSS"]
attack6 = dt[dt['Label']=="SQL Injection"]
normal = dt[dt['Label']=="Benign"]

print(attack1['Label'].value_counts())
print(attack2['Label'].value_counts())
print(attack3['Label'].value_counts())
print(attack4['Label'].value_counts())
print(attack5['Label'].value_counts())
print(attack6['Label'].value_counts())
print(normal['Label'].value_counts())

print("Label converting...")
# Convert label
label = {'DoS attacks-SlowHTTPTest': 1}
attack1['Label'] = attack1['Label'].map(label).fillna(attack1['Label'])

label2 = {'DoS attacks-Slowloris': 1}
attack2['Label'] = attack2['Label'].map(label2).fillna(attack2['Label'])

label3 = {'DDOS attack-LOIC-UDP': 1}
attack3['Label'] = attack3['Label'].map(label3).fillna(attack3['Label'])

label4 = {'Brute Force -Web': 1}
attack4['Label'] = attack4['Label'].map(label4).fillna(attack4['Label'])

label5 = {'Brute Force -XSS': 1}
attack5['Label'] = attack5['Label'].map(label5).fillna(attack5['Label'])

label6 = {'SQL Injection': 1}
attack6['Label'] = attack6['Label'].map(label6).fillna(attack6['Label'])

label7 = {'Benign': 0}
normal['Label'] = normal['Label'].map(label7).fillna(normal['Label'])

# Merge data
#DoS attacks-SlowHTTPTest
data1 = pd.concat([attack1,normal],ignore_index = True,axis=0)
data_train1 = shuffle(data1).reset_index(drop=True)

#DoS attacks-Slowloris
data2 = pd.concat([attack2,normal],ignore_index = True,axis=0)
data_train2 = shuffle(data2).reset_index(drop=True)

#DDOS attack-LOIC-UDP
data3 = pd.concat([attack3,normal],ignore_index = True,axis=0)
data_train3 = shuffle(data3).reset_index(drop=True)

#Brute Force -Web
data4 = pd.concat([attack4,normal],ignore_index = True,axis=0)
data_train4 = shuffle(data4).reset_index(drop=True)

#Brute Force -XSS
data5 = pd.concat([attack5,normal],ignore_index = True,axis=0)
data_train5 = shuffle(data5).reset_index(drop=True)

#SQL Injection
data6 = pd.concat([attack6,normal],ignore_index = True,axis=0)
data_train6 = shuffle(data6).reset_index(drop=True)

# Save file
print("-----------------------------------------")
print("DoS attacks-SlowHTTPTest Dataset:")
print(data_train1['Label'].value_counts())
print("-----------------------------------------")
data_train1.to_csv("../dataset/CIC-2018/traindoshttp.csv", index=False, header = True)
print("Saved to file train 1.csv")

print("-----------------------------------------")
print("DoS attacks-Slowloris Dataset:")
print(data_train2['Label'].value_counts())
print("-----------------------------------------")
data_train2.to_csv("../dataset/CIC-2018/traindosloris.csv", index=False, header = True)
print("Saved to file train 2.csv")

print("-----------------------------------------")
print("DDOS attack-LOIC-UDP Dataset:")
print(data_train3['Label'].value_counts())
print("-----------------------------------------")
data_train3.to_csv("../dataset/CIC-2018/trainddosudp.csv", index=False, header = True)
print("Saved to file train 3.csv")


print("-----------------------------------------")
print("Brute Force -Web Dataset:")
print(data_train4['Label'].value_counts())
print("-----------------------------------------")
data_train4.to_csv("../dataset/CIC-2018/trainbruteweb.csv", index=False, header = True)
print("Saved to file train 4.csv")

print("-----------------------------------------")
print("Brute Force -XSS Dataset:")
print(data_train5['Label'].value_counts())
print("-----------------------------------------")
data_train5.to_csv("../dataset/CIC-2018/trainbrutexss.csv", index=False, header = True)
print("Saved to file train 5.csv")

print("-----------------------------------------")
print("SQL Injection Dataset:")
print(data_train6['Label'].value_counts())
print("-----------------------------------------")
data_train6.to_csv("../dataset/CIC-2018/trainsql.csv", index=False, header = True)
print("Saved to file train 6.csv")
print("-----------------------------------------")

import numpy as np
import pandas as pd
from sklearn.utils import shuffle

dt1 = pd.read_csv("../dataset/CIC-2018/test.csv")

# Random data attack and background 
attack1 = dt1[dt1['Label']=="DoS attacks-SlowHTTPTest"]
attack2 = dt1[dt1['Label']=="DoS attacks-Slowloris"]
attack3 = dt1[dt1['Label']=="DDOS attack-LOIC-UDP"]
attack4 = dt1[dt1['Label']=="Brute Force -Web"]
attack5 = dt1[dt1['Label']=="Brute Force -XSS"]
attack6 = dt1[dt1['Label']=="SQL Injection"]
normal = dt1[dt1['Label']=="Benign"]

print("Label converting...")
# Convert label
label = {'DoS attacks-SlowHTTPTest': 1}
attack1['Label'] = attack1['Label'].map(label).fillna(attack1['Label'])

label2 = {'DoS attacks-Slowloris': 1}
attack2['Label'] = attack2['Label'].map(label2).fillna(attack2['Label'])

label3 = {'DDOS attack-LOIC-UDP': 1}
attack3['Label'] = attack3['Label'].map(label3).fillna(attack3['Label'])

label4 = {'Brute Force -Web': 1}
attack4['Label'] = attack4['Label'].map(label4).fillna(attack4['Label'])

label5 = {'Brute Force -XSS': 1}
attack5['Label'] = attack5['Label'].map(label5).fillna(attack5['Label'])

label6 = {'SQL Injection': 1}
attack6['Label'] = attack6['Label'].map(label6).fillna(attack6['Label'])

label7 = {'Benign': 0}
normal['Label'] = normal['Label'].map(label7).fillna(normal['Label'])

# Merge data
#DoS attacks-SlowHTTPTest
data1 = pd.concat([attack1,normal],ignore_index = True,axis=0)
data_train1 = shuffle(data1).reset_index(drop=True)

#DoS attacks-Slowloris
data2 = pd.concat([attack2,normal],ignore_index = True,axis=0)
data_train2 = shuffle(data2).reset_index(drop=True)

#DDOS attack-LOIC-UDP
data3 = pd.concat([attack3,normal],ignore_index = True,axis=0)
data_train3 = shuffle(data3).reset_index(drop=True)

#Brute Force -Web
data4 = pd.concat([attack4,normal],ignore_index = True,axis=0)
data_train4 = shuffle(data4).reset_index(drop=True)

#Brute Force -XSS
data5 = pd.concat([attack5,normal],ignore_index = True,axis=0)
data_train5 = shuffle(data5).reset_index(drop=True)

#SQL Injection
data6 = pd.concat([attack6,normal],ignore_index = True,axis=0)
data_train6 = shuffle(data6).reset_index(drop=True)

# Save file
print("-----------------------------------------")
print("DoS attacks-SlowHTTPTest Dataset:")
print(data_train1['Label'].value_counts())
print("-----------------------------------------")
data_train1.to_csv("../dataset/CIC-2018/testdoshttp.csv", index=False, header = True)
print("Saved to file test 1.csv")

print("-----------------------------------------")
print("DoS attacks-Slowloris Dataset:")
print(data_train2['Label'].value_counts())
print("-----------------------------------------")
data_train2.to_csv("../dataset/CIC-2018/testdosloris.csv", index=False, header = True)
print("Saved to file test 2.csv")

print("-----------------------------------------")
print("DDOS attack-LOIC-UDP Dataset:")
print(data_train3['Label'].value_counts())
print("-----------------------------------------")
data_train3.to_csv("../dataset/CIC-2018/testddosudp.csv", index=False, header = True)
print("Saved to file test 3.csv")


print("-----------------------------------------")
print("Brute Force -Web Dataset:")
print(data_train4['Label'].value_counts())
print("-----------------------------------------")
data_train4.to_csv("../dataset/CIC-2018/testbruteweb.csv", index=False, header = True)
print("Saved to file test 4.csv")

print("-----------------------------------------")
print("Brute Force -XSS Dataset:")
print(data_train5['Label'].value_counts())
print("-----------------------------------------")
data_train5.to_csv("../dataset/CIC-2018/testbrutexss.csv", index=False, header = True)
print("Saved to file test 5.csv")

print("-----------------------------------------")
print("SQL Injection Dataset:")
print(data_train6['Label'].value_counts())
print("-----------------------------------------")
data_train6.to_csv("../dataset/CIC-2018/testsql.csv", index=False, header = True)
print("Saved to file test 6.csv")
print("-----------------------------------------")

import numpy as np
import pandas as pd
from sklearn.utils import shuffle
import warnings
from fastprogress import fastprogress
warnings.filterwarnings("ignore")

df = pd.read_csv("../dataset/CIC-2018/original.csv")
dt = pd.read_csv("../dataset/CIC-2018/train_ori.csv")
dt1 = pd.read_csv("../dataset/CIC-2018/test.csv")

print(df['Label'].value_counts())
print(dt['Label'].value_counts())
print(dt1['Label'].value_counts())
