==> CIC/1_Prepare.ipynb <==
# Read original data
import dask.dataframe as dd
import numpy as np
import pandas as pd
from sklearn.utils import shuffle

dt = dd.read_csv("../dataset/CIC-2018/trainddosudp.csv")
print("Original label values: \n")
print(dt['Label'].value_counts().compute())

# Read original data
import dask.dataframe as dd
import numpy as np
import pandas as pd
from sklearn.utils import shuffle

dt = dd.read_csv("../dataset/CIC-2018/data_balanced.csv")
print("Original label values: \n")
print(dt['Label'].value_counts().compute())
print("-----------------------------------------")
# Random data dos and background 
normal = dt[dt['Label']=="Benign"].compute()
a1 = dt[dt['Label']=="Bot"].compute()
a2 = dt[dt['Label']=="DDOS attack-HOIC"].compute()
a3 = dt[dt['Label']=="DoS attacks-GoldenEye"].compute()
a4 = dt[dt['Label']=="DoS attacks-Hulk"].compute()
a5 = dt[dt['Label']=="Infilteration"].compute()
a6 = dt[dt['Label']=="DoS attacks-SlowHTTPTest"].compute()
a7 = dt[dt['Label']=="DoS attacks-Slowloris"].compute()
a8 = dt[dt['Label']=="DDOS attack-LOIC-UDP"].compute()

# Merge data
data = pd.concat([normal,a1,a2,a3,a4,a5,a6,a7,a8],ignore_index = True,axis=0)
data_train = shuffle(data).reset_index(drop=True)

print("-----------------------------------------")
print(data_train['Label'].value_counts())

print("-----------------------------------------")
data_train.to_csv("../dataset/CIC-2018/data_balanced_1.csv", index=False, header = True)
print("Saved to file ok!!!")
