==> CIC/3_CIC_merge_each.ipynb <==
import pandas as pd 
import numpy as np  
from sklearn.utils import shuffle

dt11 = pd.read_csv("../dataset/CIC-2018/traindoshttp.csv")
dt12 = pd.read_csv("../dataset/CIC-2018/generated_doshttp.csv")
dt21 = pd.read_csv("../dataset/CIC-2018/traindosloris.csv")
dt22 = pd.read_csv("../dataset/CIC-2018/generated_dosloris.csv")
dt31 = pd.read_csv("../dataset/CIC-2018/trainddosudp.csv")
dt32 = pd.read_csv("../dataset/CIC-2018/generated_ddosudp.csv")

print(dt11['Label'].value_counts())
print(dt21['Label'].value_counts())
print(dt31['Label'].value_counts())

a1 = dt12.sample(n=4460, replace=False, random_state=1) 
a2 = dt22.sample(n=8895, replace=False, random_state=1)
a3 = dt32.sample(n=13126, replace=False, random_state=1)

# Merge data
data1 = pd.concat([dt11,a1],ignore_index = True,axis=0)
data_balanced1 = shuffle(data1).reset_index(drop=True)
data_balanced1.to_csv("../dataset/CIC-2018/balanced_doshttp.csv", index=False, header = True)

data2 = pd.concat([dt21,a2],ignore_index = True,axis=0)
data_balanced2 = shuffle(data2).reset_index(drop=True)
data_balanced2.to_csv("../dataset/CIC-2018/balanced_dosloris.csv", index=False, header = True)

data3 = pd.concat([dt31,a3],ignore_index = True,axis=0)
data_balanced3 = shuffle(data3).reset_index(drop=True)
data_balanced3.to_csv("../dataset/CIC-2018/balanced_ddosudp.csv", index=False, header = True)

print("Balanced dataset:")
print(data_balanced1['Label'].value_counts())
print("-----------------------------------------")
print(data_balanced2['Label'].value_counts())
print("-----------------------------------------")
print(data_balanced3['Label'].value_counts())
print("-----------------------------------------")

import pandas as pd 
import numpy as np  
from sklearn.utils import shuffle

dt11 = pd.read_csv("../dataset/CIC-2018/trainbruteweb.csv")
dt12 = pd.read_csv("../dataset/CIC-2018/generated_bruteweb.csv")
dt21 = pd.read_csv("../dataset/CIC-2018/trainbrutexss.csv")
dt22 = pd.read_csv("../dataset/CIC-2018/generated_brutexss.csv")
dt31 = pd.read_csv("../dataset/CIC-2018/trainsql.csv")
dt32 = pd.read_csv("../dataset/CIC-2018/generated_sql.csv")

print(dt11['Label'].value_counts())
print(dt21['Label'].value_counts())
print(dt31['Label'].value_counts())

a1 = dt12.sample(n=13826, replace=False, random_state=1) 
a2 = dt22.sample(n=13924, replace=False, random_state=1)
a3 = dt32.sample(n=13981, replace=False, random_state=1)

# Merge data
data1 = pd.concat([dt11,a1],ignore_index = True,axis=0)
data_balanced1 = shuffle(data1).reset_index(drop=True)
data_balanced1.to_csv("../dataset/CIC-2018/balanced_bruteweb.csv", index=False, header = True)

data2 = pd.concat([dt21,a2],ignore_index = True,axis=0)
data_balanced2 = shuffle(data2).reset_index(drop=True)
data_balanced2.to_csv("../dataset/CIC-2018/balanced_brutexss.csv", index=False, header = True)

data3 = pd.concat([dt31,a3],ignore_index = True,axis=0)
data_balanced3 = shuffle(data3).reset_index(drop=True)
data_balanced3.to_csv("../dataset/CIC-2018/balanced_sql.csv", index=False, header = True)

print("Balanced dataset:")
print(data_balanced1['Label'].value_counts())
print("-----------------------------------------")
print(data_balanced2['Label'].value_counts())
print("-----------------------------------------")
print(data_balanced3['Label'].value_counts())
print("-----------------------------------------")
