==> CIC/4_CIC_Autogluon.ipynb <==
import pandas as pd 
import numpy as np 
import sys
from sklearn.preprocessing import MinMaxScaler 
from autogluon.tabular import TabularDataset, TabularPredictor
import warnings
from fastprogress import fastprogress
warnings.filterwarnings("ignore")

#Load a tabular & test dataset:
train_data = TabularDataset('../dataset/CIC-2018/train_ori.csv')
test_data = TabularDataset('../dataset/CIC-2018/test.csv')

print(train_data['Label'].value_counts())
print(test_data['Label'].value_counts())


print("Fit classification models predicting:")
predictor = TabularPredictor(label='Label').fit(train_data)  # Fit models for 12000s # time_limit=12000, presets='best_quality'

print("Evaluate predictions on test data:")
leaderboard = predictor.leaderboard(test_data)

print("The best model:")
print(predictor.get_model_best())
print("------------------------------------------------")
print(predictor.evaluate(test_data))

metric=['accuracy', 'precision_weighted', 'f1_weighted', 'recall_weighted']
predictor.leaderboard(test_data, extra_metrics=metric, silent=True)
