==> CIC/4_merge_full.ipynb <==
# Read original data
import numpy as np
import pandas as pd
from sklearn.utils import shuffle

dt = pd.read_csv("../dataset/CIC-2018/train_ori.csv")
dt1 = pd.read_csv("../dataset/CIC-2018/generated_doshttp.csv")
dt2 = pd.read_csv("../dataset/CIC-2018/generated_dosloris.csv")
dt3 = pd.read_csv("../dataset/CIC-2018/generated_ddosudp.csv")
#dt4 = pd.read_csv("../dataset/CIC-2018/generated_bruteweb.csv")
#dt5 = pd.read_csv("../dataset/CIC-2018/generated_brutexss.csv")
#dt6 = pd.read_csv("../dataset/CIC-2018/generated_sql.csv")

# Convert label
print("Label converting...")
label = {1: 'DoS attacks-SlowHTTPTest'}
dt1['Label'] = dt1['Label'].map(label).fillna(dt1['Label'])

label2 = {1: 'DoS attacks-Slowloris'}
dt2['Label'] = dt2['Label'].map(label2).fillna(dt2['Label'])

label3 = {1: 'DDOS attack-LOIC-UDP'}
dt3['Label'] = dt3['Label'].map(label3).fillna(dt3['Label'])

label4 = {1: 'Brute Force -Web'}
dt4['Label'] = dt4['Label'].map(label4).fillna(dt4['Label'])

label5 = {1: 'Brute Force -XSS'}
dt5['Label'] = dt5['Label'].map(label5).fillna(dt5['Label'])

label6 = {1: 'SQL Injection'}
dt6['Label'] = dt6['Label'].map(label6).fillna(dt6['Label'])

print('Completed!')

print(dt['Label'].value_counts())
print(dt1['Label'].value_counts())
print(dt2['Label'].value_counts())
print(dt3['Label'].value_counts())
#print(dt4['Label'].value_counts())
#print(dt5['Label'].value_counts())
#print(dt6['Label'].value_counts())

a1 = dt1.sample(n=4459, replace=False, random_state=1) #doshttp
a2 = dt2.sample(n=8895, replace=False, random_state=1)
a3 = dt3.sample(n=13126, replace=False, random_state=1)
#a4 = dt4.sample(n=13826, replace=False, random_state=1)
#a5 = dt5.sample(n=13924, replace=False, random_state=1)
#a6 = dt6.sample(n=13981, replace=False, random_state=1)

# Merge data
data = pd.concat([dt,a1,a2,a3],ignore_index = True,axis=0)
data_balanced = shuffle(data).reset_index(drop=True)
data_balanced.to_csv("../dataset/CIC-2018/train.csv", index=False, header = True)
print("Original dataset:")
print(dt['Label'].value_counts())
print("-----------------------------------------")
print("Balanced dataset:")
print(data_balanced['Label'].value_counts())
print("-----------------------------------------")
